refactor: rename setup.js to testSetup.js
